requirements.txt
langchain
python-dotenv
ipykernel
langchain-openai
langchain_community
pypdf
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
duckdb
pandas
beautifulsoup4
faiss-cpu
langchain_groq
langchain_core
fastapi
uvicorn
langserve
langchain_huggingface
langchain-chroma
ipywidgets
streamlit
openai